feat: multimodule -> news-service. size of varchar in *.sql
