+++ news-service/src/main/java/ru/clevertec/newsservice/NewsServiceApplication.java
@@ -1,0 +1,13 @@
+package ru.clevertec.newsservice;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class NewsServiceApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(NewsServiceApplication.class, args);
+    }
+
+}
