+++ news-service/src/main/java/ru/clevertec/newsservice/entity/Comment.java
@@ -1,39 +1,40 @@
-package ru.clevertec.paramonov.news.entity;
+package ru.clevertec.newsservice.entity;
+
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.OneToMany;
+import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import java.time.LocalDateTime;
-import java.util.List;
 
 @Getter
 @Setter
 @NoArgsConstructor
 @Entity
-@Table(name = "news")
-public class News implements BaseEntity<Long>{
+@Table(name = "comment")
+public class Comment implements BaseEntity<Long> {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "news_id")
+    @Column(name = "comment_id")
     private Long id;
 
     @Column(name = "creation_time")
     private LocalDateTime time;
 
-    private String title;
-
     @Column(name = "subject")
     private String text;
 
-    @OneToMany
-    private List<Comment> comments;
+    @ManyToOne
+    private User username;
+
+    @ManyToOne
+    private News newsId;
 }
