+++ news-service/src/main/java/ru/clevertec/newsservice/entity/News.java
@@ -1,0 +1,39 @@
+package ru.clevertec.newsservice.entity;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.OneToMany;
+import jakarta.persistence.Table;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+import java.time.LocalDateTime;
+import java.util.List;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@Entity
+@Table(name = "news")
+public class News implements BaseEntity<Long>{
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "news_id")
+    private Long id;
+
+    @Column(name = "creation_time")
+    private LocalDateTime time;
+
+    private String title;
+
+    @Column(name = "subject")
+    private String text;
+
+    @OneToMany
+    private List<Comment> comments;
+}
