+++ news-service/src/main/resources/application.yml
@@ -9,8 +9,3 @@ spring:
     database-platform: org.hibernate.dialect.PostgreSQLDialect
     hibernate:
       ddl-auto: update
-#  sql:
-#    init:
-#      mode: always
-#      schema-locations: classpath*:classpath*:database/initDB.sql
-#      data-locations: classpath*:classpath*:database/populateDB.sql
