+++ news-service/src/main/resources/db/changelog/db.changelog-1.0-29.04.sql
@@ -7,8 +7,8 @@ CREATE TABLE IF NOT EXISTS news
 (
     news_id       BIGSERIAL PRIMARY KEY,
     creation_time TIMESTAMP NOT NULL,
-    subject       VARCHAR   NOT NULL,
-    title         VARCHAR   NOT NULL
+    subject       VARCHAR(1000)   NOT NULL,
+    title         VARCHAR(128)   NOT NULL
 );
 --rollback DROP TABLE news CASCADE
 
@@ -17,7 +17,7 @@ CREATE TABLE IF NOT EXISTS comment
 (
     comment_id    BIGSERIAL PRIMARY KEY,
     creation_time TIMESTAMP NOT NULL,
-    subject       VARCHAR   NOT NULL
+    subject       VARCHAR(256)   NOT NULL
 );
 --rollback DROP TABLE comment CASCADE
 
@@ -25,8 +25,8 @@ CREATE TABLE IF NOT EXISTS comment
 CREATE TABLE IF NOT EXISTS users
 (
     user_id  BIGSERIAL PRIMARY KEY,
-    username VARCHAR NOT NULL,
-    password VARCHAR NOT NULL,
-    role     VARCHAR NOT NULL
+    username VARCHAR(64) NOT NULL,
+    password VARCHAR(256) NOT NULL,
+    role     VARCHAR(32) NOT NULL
 );
 --rollback DROP TABLE users CASCADE
