+++ news-service/src/main/resources/db/changelog/db.changelog-2.0-29.04.sql
