+++ news-service/src/main/resources/db/changelog/db.changelog-master.yaml
