--- src/main/java/ru/clevertec/paramonov/news/OwnTwitterApplication.java
@@ -1,13 +1,0 @@
-package ru.clevertec.paramonov.news;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class OwnTwitterApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(OwnTwitterApplication.class, args);
-	}
-
-}
