--- src/main/java/ru/clevertec/paramonov/news/entity/User.java
@@ -1,31 +1,0 @@
-package ru.clevertec.paramonov.news.entity;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import ru.clevertec.paramonov.news.model.Role;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@Entity
-@Table(name = "users")
-public class User implements BaseEntity<Long> {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "user_id")
-    private Long id;
-
-    private String username;
-
-    private char[] password;
-
-    private Role role;
-}
