--- src/main/resources/populateDB.sql
@@ -1,10 +1,0 @@
-INSERT INTO users
-VALUES (1, 'Antonissimo', 'Uri', 'Antonov'),
-       (2, 'Birobidgan', 'Dalekie', 'Zemli'),
-       (3, 'moska', 'Volod', 'Joke');
-
-INSERT INTO news
-VALUES (1, 123123, 123, 1),
-       (2, 23222, 321, 1),
-       (3, 3333333333, 111, 2),
-       (4, 4444444, 999, 3)
